Add make targets for assembler output

+++ fib.h
@@ -5,7 +5,7 @@ typedef unsigned int input_t;
 #ifdef FIB128
 typedef unsigned __int128 output_t;
 #else
-typedef unsigned long long output_t;
+typedef unsigned long output_t;
 #endif
 
 output_t fibarray(input_t n);
